% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca.R
\name{asca}
\alias{asca}
\title{Title}
\usage{
asca(formula, data, subset, weights, na.action, family, pca.in = FALSE)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{subset}{}

\item{weights}{}

\item{na.action}{}

\item{family}{}

\item{pca.in}{}
}
\value{

}
\description{
Title
}
\examples{
dataset1   <- data.frame(y=I(matrix(rnorm(24*10),ncol=10)), x=factor(c(rep(2,8),rep(1,8),rep(0,8))), z=factor(rep(c(1,0),12)), w=rnorm(24))
mod <- asca(y~x+z, data=dataset1)
mod <- asca(y~x+z, data=dataset1, pca.in=0)
mod <- asca(y~x+z+w, data=dataset1, family="gaussian")
mod <- asca(y~x+(1|z), data=dataset1)
mod <- asca(y~x+(1|z)+w, data=dataset1, family="gaussian")
}
