% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{import}
\name{basic}
\alias{basic}
\alias{pca}
\alias{pcr}
\alias{plsr}
\alias{cca}
\alias{gsvd}
\alias{ifa}
\alias{reexports}
\title{Single- and Two-Block Methods}
\usage{
pca(X, scale = FALSE, ncomp = 1, ...)

cca(X)

ifa(X, ncomp = 1, scale = FALSE, verbose = FALSE, ...)

gsvd(X)
}
\description{
This documentation covers a range of single- and two-block methods. In particular:
\itemize{
\item PCA - Principal Component Analysis (\code{pca})
\item PCR - Principal Component Regression (\code{pcr})
\item PLSR - Partial Least Squares Regression (\code{plsr})
\item CCA - Canonical Correlation Analysis (\code{cca})
\item IFA - Interbattery Factor Analysis (\code{ifa})
\item GSVD - Generalized SVD (\code{gsvd})
}
}
\examples{
data(potato)
X <- potato$Chemical
y <- potato$Sensory[,1,drop=FALSE]

pca.pot  <- pca(X, ncomp = 2)
pcr.pot  <- pcr(y ~ X, ncomp = 2)
pls.pot  <- plsr(y ~ X, ncomp = 2)
cca.pot  <- cca(potato[1:2])
ifa.pot  <- ifa(potato[1:2])
gsvd.pot <- gsvd(lapply(potato[3:4], t))

}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pls}{\code{\link[pls:mvr]{pcr}}, \code{\link[pls:mvr]{plsr}}}
}}

