% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{candies}
\alias{candies}
\title{Sensory assessment of candies.}
\format{
A data.frame having 165 rows and 3 variables:
\describe{
\item{assessment}{Matrix of sensory attributes}
\item{assessor}{Factor of assessors}
\item{candy}{Factor of candies}
}
}
\usage{
data(candies)
}
\description{
A dataset containing 9 sensory attributes for 5 candies assessed
by 11 trained assessors.
}
\references{
Luciano G, Næs T. Interpreting sensory data by combining principal
component analysis and analysis of variance. Food Qual Prefer. 2009;20(3):167-175.
}
