% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{cca}
\alias{cca}
\title{Canonical Correlation Analysis - CCA}
\usage{
cca(X)
}
\arguments{
\item{X}{\code{matrix} of input data.}
}
\value{
\code{multiblock} object with associated with printing, scores, loadings, ...
}
\description{
Canonical Correlation Analysis - CCA
}
\references{
Hotelling, H. (1936) Relations between two sets of variates. Biometrika, 28, 321–377.
}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
