% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compnames}
\alias{compnames}
\title{Vector of component names}
\usage{
compnames(object, comps, explvar = FALSE, ...)
}
\arguments{
\item{object}{An object fitted using the multiblock package.}

\item{comps}{\code{integer} vector of components.}

\item{explvar}{\code{logical} indicating if explained variances should be included.}

\item{...}{Unused}
}
\value{
\code{character} vector of component names.
}
\description{
Copy of \code{compnames} from \code{pls} package to work with
\code{multiblock} objects.
}
