% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{disco}
\alias{disco}
\title{Distinctive and Common Components with SCA - DISCO}
\usage{
disco(X, ncomp = 2, ...)
}
\arguments{
\item{X}{\code{list} of input blocks.}

\item{ncomp}{\code{integer} number of components to extract.}

\item{...}{additional arguments (not used).}
}
\description{
Distinctive and Common Components with SCA - DISCO
}
\examples{
data(potato)
potList <- as.list(potato[c(1,2,9)])
pot.disco  <- disco(potList)
plot(scores(pot.disco), labels="names")

}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
