% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dummycode}
\alias{dummycode}
\title{Dummy-coding of a single vector}
\usage{
dummycode(Y, contrast = "contr.sum", drop = TRUE)
}
\arguments{
\item{Y}{\code{vector} to dummy code.}

\item{contrast}{Contrast type, default = "contr.sum".}

\item{drop}{\code{logical} indicating if one level should be dropped (default = TRUE).}
}
\value{
\code{matrix} made by dummy-coding the input vector.
}
\description{
Flexible dummy-coding allowing for all R's built-in types of contrasts
and optional dropping of a factor level to reduce rank defficiency probability.
}
\examples{
vec <- c("a","a","b","b","c","c")
dummycode(vec)
}
