% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{emulsions}
\alias{emulsions}
\title{Spectroscopic analysis of emulsions}
\format{
A data.frame having 69 rows and 3 variables:
\describe{
\item{Raman}{Matrix of Raman measurements}
\item{NIR}{Matrix of Near-infrared measurements}
\item{PUFA}{Matrix of polyunsaturated fatty acid contents as a percentage
of the sample and as a percentage of the total fat content.}
}
}
\usage{
data(emulsions)
}
\description{
A dataset containing NIR and Raman measurements of emulsions in
a food model system of fat, water and protein mixtures.
}
\references{
N.K. Afseth, V.H. Segtnan, B.J. Marquardt, and J.P. Wold. Raman
and Near-Infrared Spectroscopy for Quantification of Fat Composition in a
Complex Food Model System. Applied Spectroscopy 2005;59(11):1324-1332.
}
