% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{gpa}
\alias{gpa}
\title{Generalized Procrustes Analysis - GPA}
\usage{
gpa(X, graph = FALSE, ...)
}
\arguments{
\item{X}{\code{list} of input blocks.}

\item{graph}{\code{logical} indicating if decomposition should be plotted.}

\item{...}{additional arguments for RGCCA approach.}
}
\value{
\code{multiblock} object including relevant scores and loadings.
}
\description{
Generalized Procrustes Analysis - GPA
}
\examples{
data(potato)
potList <- as.list(potato[c(1,2,9)])
pot.gpa    <- gpa(potList)
plot(scores(pot.gpa), labels="names")

}
\references{
Gower, J. C. (1975). Generalized procrustes analysis. Psychometrika. 40: 33–51.
}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
