% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{hogsvd}
\alias{hogsvd}
\title{Higher Order Generalized SVD - HOGSVD}
\usage{
hogsvd(X)
}
\arguments{
\item{X}{\code{list} of input blocks.}
}
\description{
Higher Order Generalized SVD - HOGSVD
}
\examples{
data(candies)
candyList <- lapply(1:nlevels(candies$candy),function(x)candies$assessment[candies$candy==x,])
can.hogsvd <- hogsvd(candyList)
plot(scores(can.hogsvd, block=1), labels="names")

}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
