% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ifa}
\alias{ifa}
\title{Inter-battery Factor Analysis - IFA}
\usage{
ifa(X, ncomp = 1, scale = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{X}{\code{matrix} of input data.}

\item{ncomp}{\code{integer} number of principal components to return.}

\item{scale}{\code{logical} indicating if variables should be standardised (default=FALSE).}

\item{verbose}{\code{logical} indicating if intermediate results should be printed.}

\item{...}{additional arguments to \code{pls:pcr}.}
}
\value{
\code{multiblock} object with associated with printing, scores, loadings, ...
}
\description{
Inter-battery Factor Analysis - IFA
}
\references{
Tucker, L. R. (1958). An inter-battery method of factor analysis. Psychometrika, 23(2), 111-136.
}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
