% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{jive}
\alias{jive}
\title{Joint and Individual Variation Explained - JIVE}
\usage{
jive(X, ...)
}
\arguments{
\item{X}{\code{list} of input blocks.}

\item{...}{additional arguments for \code{r.jive::jive}.}
}
\description{
Joint and Individual Variation Explained - JIVE
}
\examples{
data(candies)
candyList <- lapply(1:nlevels(candies$candy),function(x)candies$assessment[candies$candy==x,])
can.jive  <- jive(candyList)
summary(can.jive)

}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
