% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpls.R
\name{lpls}
\alias{lpls}
\title{L-PLS regression}
\usage{
lpls(
  X1,
  X2,
  X3,
  ncomp = 2,
  doublecenter = TRUE,
  scale = c(FALSE, FALSE, FALSE),
  type = c("exo"),
  impute = FALSE,
  niter = 25,
  subsetX2 = NULL,
  subsetX3 = NULL,
  ...
)
}
\arguments{
\item{X1}{\code{matrix} of size IxN (middle matrix)}

\item{X2}{\code{matrix} of size IxJ (left matrix)}

\item{X3}{\code{matrix} of size KxN (top matrix)}

\item{ncomp}{number of L-PLS components}

\item{doublecenter}{\code{logical} indicating if centering should be done both ways for X1 (default=TRUE)}

\item{scale}{\code{logical vector} of length three indicating if each of the matrices should be autoscaled.}

\item{type}{\code{character} indicating type of L-PLS ("exo"=default, "exo_ort" or "endo")}

\item{impute}{\code{logical} indicating if SVD-based imputation of missing data is required.}

\item{niter}{\code{numeric} giving number of iterations in component extraction loop.}

\item{subsetX2}{\code{vector} defining optional sub-setting of X2 data.}

\item{subsetX3}{\code{vector} defining optional sub-setting of X3 data.}

\item{...}{Additional arguments, not used.}
}
\value{

}
\description{
Simultaneous decomposition of three blocks connected in an L pattern. Two versions of L-PLS
are available: exo- and endo-L-PLS. See references for details.
}
\details{
See chapter Three-block data modeling by endo- and exo-L-PLS regression by Sæbø, S., Martens,
M and Martens, M. for details on the endo- and exo-L-PLS algorithms. Endo-L-PLS is an inward
regression of X2 and X3 on X1, whereas exo-L-PLS is an outward regression of X1 on X2 and X3.
The \code{exo_ort} algorithm returns orthogonal scores and should be chosen for visual
exploration in correlation loading plots. If exo-L-PLS with prediction is the main purpose
of the model then the non-orthogonal \code{exo} type L-PLS should be chosen for which the
predict function has prediction implemented. \cr
\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} ______N\Sexpr{"\u200B"} }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |}\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |}\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} X3\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |}\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |}\cr
\code{\Sexpr{"\u200B"} K|_______|}\cr
\code{\Sexpr{"\u200B"} }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} ______N\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} ________J\Sexpr{"\u200B"} }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |}\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |}\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} X1\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} X2\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |}\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |}\cr
\code{\Sexpr{"\u200B"} I|_______|\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"}I|_________|}\cr
}
\examples{
# Simulate data set
sim <- lplsData(I = 30, N = 20, J = 5, K = 6, ncomp = 2)
X1  <- sim$X1; X2 <- sim$X2; X3 <- sim$X3
lp  <- lpls(X1,X2,X3) # exo-L-PLS
}
\references{
Sæbø, S., Martens, M. and Martens H. (2010) Three-block data modeling by endo- and
exo-LPLS regression. In Handbook of Partial Least Squares: Concepts, Methods and Applications.
Esposito Vinzi, V.; Chin, W.W.; Henseler, J.; Wang, H. (Eds.). Springer.
}
\author{
Solve Sæbø (adapted by Kristian Hovde Liland)
}
