% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lplsData.R
\name{lplsData}
\alias{lplsData}
\title{L-PLS data simulation for exo-type analysis}
\usage{
lplsData(I = 30, N = 20, J = 5, K = 6, ncomp = 2)
}
\arguments{
\item{I}{\code{numeric} number of rows of X1 and X2}

\item{N}{\code{numeric} number of columns in X1 and X3}

\item{J}{\code{numeric} number of columns in X2}

\item{K}{\code{numeric} number of rows in X3}

\item{ncomp}{\code{numeric} number of latent components}
}
\value{

}
\description{
L-PLS data simulation for exo-type analysis
}
\author{
Solve Sæbø (adapted by Kristian Hovde Liland)
}
