% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised.R
\name{mbpls}
\alias{mbpls}
\title{Multiblock Partial Least Squares - MB-PLS}
\usage{
mbpls(X, Y, ncomp = 1, scale = FALSE, ...)
}
\arguments{
\item{X}{\code{list} of input blocks.}

\item{Y}{\code{matrix} of responses.}

\item{ncomp}{\code{integer} number of PLS components.}

\item{scale}{\code{logical} for autoscaling inputs (default = FALSE).}

\item{...}{additional arguments to pls::plsr.}
}
\value{
\code{mbpls} object containing the underlying \code{pls} object, with all its result and plot possibilities plus block-wise loadings, loading weights and scores.
}
\description{
Multiblock Partial Least Squares - MB-PLS
}
\examples{
data(potato)
mb <- mbpls(potato[c('Chemical','Compression')], potato[['Sensory']], ncomp = 5)
print(mb)
scoreplot(mb, labels="names") # Exploiting mvr object structure from pls package
}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
