% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised.R
\name{mbrda}
\alias{mbrda}
\title{Multiblock Redundancy Analysis - mbRDA}
\usage{
mbrda(X, Y, ncomp = 1, ...)
}
\arguments{
\item{X}{\code{list} of input blocks.}

\item{Y}{\code{matrix} of responses.}

\item{ncomp}{\code{integer} number of PLS components.}

\item{...}{additional arguments to ade4::mbpcaiv.}
}
\value{
\code{mbrda,mvr} object containing elements corresponding to a \code{pls} object, with all its result and plot possibilities plus block-wise loadings, loading weights and scores.
}
\description{
Multiblock Redundancy Analysis - mbRDA
}
\examples{
# Convert data.frame with AsIs objects to list of matrices
data(potato)
potatoList <- lapply(potato, unclass)

mbr <- mbrda(potatoList[c('Chemical','Compression')], potatoList[['Sensory']], ncomp = 10)
print(mbr)
scoreplot(mbr) # Exploiting mvr object structure from pls package
}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
