% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{mcoa}
\alias{mcoa}
\title{Multiple Co-Inertia Analysis - MCOA}
\usage{
mcoa(X, ncomp = 2, scale = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{X}{\code{list} of input blocks.}

\item{ncomp}{\code{integer} number of components to extract.}

\item{scale}{\code{logical} indicating if variables should be scaled.}

\item{verbose}{\code{logical} indicating if diagnostic information should be printed.}

\item{...}{additional arguments for RGCCA.}
}
\description{
Multiple Co-Inertia Analysis - MCOA
}
\examples{
data(potato)
potList <- as.list(potato[c(1,2,9)])
pot.mcoa   <- mcoa(potList)
plot(scores(pot.mcoa), labels="names")

}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
