% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mcolors}
\alias{mcolors}
\title{Color palette generation from matrix of RGB values}
\usage{
mcolors(
  n,
  colmatrix = matrix(c(0, 0, 1, 1, 1, 1, 1, 0, 0), 3, 3, byrow = TRUE)
)
}
\arguments{
\item{n}{Integer number of colors to produce.}

\item{colmatrix}{A numeric \code{matrix} of three columns (R,G,B) to generate color palette from.}
}
\value{
A vector of n colors.
}
\description{
Color palette generation from matrix of RGB values
}
