% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock-package.R
\docType{package}
\name{multiblock}
\alias{multiblock}
\title{multiblock}
\description{
A collection of methods for analysis of data sets with more than two blocks of data.

\strong{Unsupervised methods:}
\itemize{
\item SCA - Simultaneous Component Analysis (\code{sca})
\item GCA - Generalized Canonical Analysis (\code{gca})
\item GPA - Generalized Procrustes Analysis (\code{gpa})
\item MFA - Multiple Factor Analysis (\code{mfa})
\item PCA-GCA (\code{pcagca})
\item DISCO - Distinctive and Common Components with SCA (\code{disco})
\item HPCA - Hierarchical Principal component analysis (\code{hpca})
\item MCOA - Multiple Co-Inertia Analysis (\code{mcoa})
\item JIVE - Joint and Individual Variation Explained (\code{jive})
\item STATIS - Structuration des Tableaux à Trois Indices de la Statistique (\code{statis})
\item HOGSVD - Higher Order Generalized SVD (\code{hogsvd})
}

\strong{Design based methods:}
\itemize{
\item ASCA - Anova Simultaneous Component Analysis (\code{asca})
}

\strong{Supervised methods:}
\itemize{
\item MB-PLS - Multiblock Partial Least Squares (\code{mbpls})
\item SO-PLS - Sequential and Orthogonalized PLS (\code{\link{sopls}})
\item ROSA - Response Oriented Sequential Alternation (\code{\link{rosa}})
\item mbRDA - Multiblock Redundancy Analysis (\code{mbrda})
}

\strong{Single- and two-block methods:}
\itemize{
\item PCA - Principal Component Analysis (\code{pca})
\item PCR - Principal Component Regression (\code{pcr})
\item PLSR - Partial Least Squares Regression (\code{plsr})
\item CCA - Canonical Correlation Analysis (\code{cca})
\item IFA - Interbattery Factor Analysis (\code{ifa})
\item GSVD - Generalized SVD (\code{gsvd})
}

\strong{Complex methods:}
\itemize{
\item L-PLS - Partial Least Squares in L configuration (\emph{lpls})
\item SO-PLS-PM - Sequential and Orthogonalised PLS Path Modeling (\emph{sopls_pm})
}
}
