% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_results.R
\name{multiblock_object}
\alias{multiblock_object}
\alias{scores.multiblock}
\alias{loadings.multiblock}
\alias{print.multiblock}
\alias{summary.multiblock}
\title{Result Functions for Supervised and Multiblock Objects}
\usage{
\method{scores}{multiblock}(object, block = 0, ...)

\method{loadings}{multiblock}(object, block = 0, ...)

\method{print}{multiblock}(x, ...)

\method{summary}{multiblock}(object, ...)
}
\arguments{
\item{object}{\code{multiblock} object.}

\item{block}{\code{integer/character} for block selection.}

\item{...}{Not implemented.}

\item{x}{\code{multiblock} object.}
}
\description{
Result Functions for Supervised and Multiblock Objects
}
\examples{
data(wine)
sc <- sca(wine[c('Smell at rest', 'View', 'Smell after shaking')], ncomp = 4)
print(sc)
plot(loadings(sc, block = 1), labels = "names", scatter = TRUE)

}
\seealso{
Overviews of available methods, \code{\link{multiblock}}, and methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
