% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{process}
\alias{process}
\alias{process_data}
\title{Cheese production process data.}
\format{
A list of matrices having 795 rows and varying numbers of variables:
\describe{
\item{A}{Process stage A}
\item{B}{Process stage B}
...
}
}
\usage{
data(process)
}
\description{
A dataset containing process data for 4 connected process (A,B,C,D) stages,
three unconnected inputs (A,B,C) and a set of endpoint qualities (E).
The stages A, B and C are influencing D and A, B, C and D are influencing E.
This can be described as a directed asyclic graph (sketched below). \cr

\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} B --- }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \ \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \ }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} --+-- \Sexpr{"\u200B"} | }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} /\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} |\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \| }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} A --- D -- E }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} /\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} / }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} C --- }\cr

Subpaths include: ADE, AE, BDE, BE, CDE, and CE, and SO-PLS-PM models model
A+B+C->D and A+B+C+D->E.
}
