% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_export.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pcr}
\alias{plsr}
\alias{MSEP}
\alias{R2}
\alias{RMSEP}
\alias{coefplot}
\alias{cvsegments}
\alias{loading.weights}
\alias{loadingplot}
\alias{loadings}
\alias{mvrValstats}
\alias{predplot}
\alias{scoreplot}
\alias{scores}
\alias{validationplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pls}{\code{\link[pls]{coefplot}}, \code{\link[pls]{cvsegments}}, \code{\link[pls:scores]{loading.weights}}, \code{\link[pls:scoreplot]{loadingplot}}, \code{\link[pls:scores]{loadings}}, \code{\link[pls:mvrVal]{MSEP}}, \code{\link[pls:mvrVal]{mvrValstats}}, \code{\link[pls:mvr]{pcr}}, \code{\link[pls:mvr]{plsr}}, \code{\link[pls]{predplot}}, \code{\link[pls:mvrVal]{R2}}, \code{\link[pls:mvrVal]{RMSEP}}, \code{\link[pls]{scoreplot}}, \code{\link[pls]{scores}}, \code{\link[pls]{validationplot}}}
}}

