% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosa.R
\name{rosa}
\alias{rosa}
\title{Response Oriented Sequential Alternation}
\usage{
rosa(
  formula,
  ncomp,
  Y.add,
  common.comp = 1,
  data,
  subset,
  na.action,
  scale = FALSE,
  weights = NULL,
  validation = c("none", "CV", "LOO"),
  internal.validation = "LOO",
  fixed.block = NULL,
  design.block = NULL,
  canonical = TRUE,
  ...
)
}
\arguments{
\item{formula}{Model formula accepting a single response (block) and predictor block names separated by + signs.}

\item{ncomp}{The maximum number of ROSA components.}

\item{Y.add}{Optional response(s) available in the data set.}

\item{common.comp}{Automatically create all combinations of common components up to length \code{common.comp} (default = 1).}

\item{data}{The data set to analyse.}

\item{subset}{Expression for subsetting the data before modelling.}

\item{na.action}{How to handle NAs (no action implemented).}

\item{scale}{Optionally scale predictor variables by their individual standard deviations.}

\item{weights}{Optional object weights.}

\item{validation}{Optional cross-validation strategy "CV" or "LOO".}

\item{fixed.block}{integer vector with block numbers for each component (0 = not fixed) or list of length <= ncomp (element length 0 = not fixed).}

\item{design.block}{integer vector containing block numbers of design blocks}

\item{canonical}{logical indicating if canonical correlation should be use when calculating loading weights (default), enabling B/W maximization, common components, etc. Alternatively (FALSE) a PLS2 strategy, e.g. for spectra response, is used.}

\item{...}{Additonal arguments for \code{cvseg} or \code{rosa.fit}}

\item{lower}{numerical indicating lower bound for powered PLS. Default = 0.5.}

\item{upper}{numerical indicating upper bound for powered PLS. Default = 0.5.}
}
\value{
An object of classes \code{rosa} and \code{mvr} having several associated printing and plotting methods.
}
\description{
Response Oriented Sequential Alternation
}
\examples{
data(potato)
mod <- rosa(Sensory[,1] ~ ., data = potato, ncomp = 10, validation = "CV", segments = 5)
summary(mod)
}
