% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosa_results.R
\name{rosa_results}
\alias{rosa_results}
\alias{predict.rosa}
\alias{rosa.classify}
\alias{coef.rosa}
\alias{print.rosa}
\alias{summary.rosa}
\alias{blockexpl}
\alias{print.rosaexpl}
\title{Result functions for ROSA models}
\usage{
\method{predict}{rosa}(
  object,
  newdata,
  ncomp = 1:object$ncomp,
  comps,
  type = c("response", "scores"),
  na.action = na.pass,
  ...
)

\method{coef}{rosa}(object, ncomp = object$ncomp, comps, intercept = FALSE, ...)

\method{print}{rosa}(x, ...)

\method{summary}{rosa}(
  object,
  what = c("all", "validation", "training"),
  digits = 4,
  print.gap = 2,
  ...
)

blockexpl(object, ncomp = object$ncomp, type = c("train", "CV"))

\method{print}{rosaexpl}(x, digits = 3, compwise = FALSE, ...)

rosa.classify(object, classes, newdata, ncomp, LQ)
}
\arguments{
\item{object}{A \code{rosa} object.}

\item{newdata}{Optional new data with the same types of predictor blocks as the ones used for fitting the object.}

\item{ncomp}{An \code{integer} giving the number of components to use apply.}

\item{comps}{An \code{integer} vector giving the exact components to apply.}

\item{type}{Character indicating which type of explained variance to compute (default = "train", alternative = "CV").}

\item{na.action}{Function determining what to do with missing values in \code{newdata}.}

\item{...}{Additional arguments. Currently not implemented.}

\item{intercept}{A \code{logical} indicating if coefficients for the intercept should be included (default = FALSE).}

\item{x}{A \code{rosa} object.}

\item{what}{A \code{character} indicating if summary should include all, validation or training.}

\item{digits}{The number of digits used for printing.}

\item{print.gap}{Gap between columns when printing.}

\item{compwise}{Logical indicating if block-wise (default/FALSE) or component-wise (TRUE) explained variance should be printed.}

\item{classes}{A \code{character} vector of class labels.}

\item{LQ}{A \code{character} indicating if 'max' (maximum score value), 'lda' or 'qda' should be used when classifying.}
}
\value{
Returns depend on method used, e.g. \code{predict.rosa} returns predicted responses
or scores depending on inputs, \code{coef.rosa} return regression coefficients, \code{blockexpl}
returns an object of class \code{rosaexpl} containing block-wise and component-wise explained variance contained in a matrix with attributes.
}
\description{
Result functions for ROSA models
}
\examples{
data(potato)
mod <- rosa(Sensory[,1] ~ ., data = potato, ncomp = 5, subset = 1:20)
testset <- potato[-(1:20),]; testset$Sensory <- NULL#testset$Sensory[,1]
predict(mod, testset, ncomp=5)
dim(coef(mod, ncomp=5)) # <variables x responses x components>
print(mod)
summary(mod)
blockexpl(mod)
print(blockexpl(mod), compwise=TRUE)

}
