% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{sca}
\alias{sca}
\title{Simultaneous Component Analysis - SCA}
\usage{
sca(X, ncomp = 2, scale = FALSE, samplelinked = "auto", ...)
}
\arguments{
\item{X}{\code{list} of input blocks.}

\item{ncomp}{\code{integer} number of components to extract.}

\item{scale}{\code{logical} indicating autoscaling of features (default = FALSE).}

\item{samplelinked}{\code{character/logical} indicating if blocks are linked by samples (TRUE) or variables (FALSE). Using 'auto' (default), this will be determined automatically.}

\item{...}{additional arguments (not used).}
}
\value{
\code{multiblock} object including relevant scores and loadings.
}
\description{
Simultaneous Component Analysis - SCA
}
\examples{
# Object linked data
data(potato)
potList <- as.list(potato[c(1,2,9)])
pot.sca    <- sca(potList)
plot(scores(pot.sca), labels="names")

# Variable linked data
data(candies)
candyList <- lapply(1:nlevels(candies$candy),function(x)candies$assessment[candies$candy==x,])
pot.sca    <- sca(candyList, samplelinked = FALSE)
pot.sca

}
\references{
Levin, J. (1966) Simultaneous factor analysis of several gramian matrices. Psychometrika, 31(3), 413–419.
}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
