% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simulated}
\alias{simulated}
\title{Data simulated to have certain characteristics.}
\format{
A list of matrices having 200 rows and 10 variables:
\describe{
\item{A}{Simulated matrix A}
\item{B}{Simulated matrix B}
...
}
}
\usage{
data(simulated)
}
\description{
A dataset containing simulated data for 4 connected events where A is the
starting point and D is the end point. This can be described as a directed
asyclic graph (sketched below, moving left->right). \cr

\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} C    }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} /^\   }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} / \Sexpr{"\u200B"}| \  }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} A--+->D }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \ \Sexpr{"\u200B"}| /  }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \|/   }\cr
\code{\Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} \Sexpr{"\u200B"} B    }\cr

Subpaths include: ABD, AD, ABCD, ACD
}
\references{
Næs, Romano, ..., Liland, In preparation.
}
