% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sopls_results.R
\name{sopls_object}
\alias{sopls_object}
\alias{predict.sopls}
\alias{sopls.classify}
\alias{coef.sopls}
\alias{print.sopls}
\alias{summary.sopls}
\alias{classify}
\alias{classify.sopls}
\alias{R2.sopls}
\alias{RMSEP.sopls}
\title{Result functions for SO-PLS models}
\usage{
\method{predict}{sopls}(
  object,
  newdata,
  ncomp = object$ncomp,
  comps = object$ncomp,
  type = c("response", "scores"),
  na.action = na.pass,
  ...
)

\method{coef}{sopls}(
  object,
  ncomp = object$ncomp,
  comps = object$ncomp,
  intercept = FALSE,
  ...
)

\method{print}{sopls}(x, ...)

\method{summary}{sopls}(
  object,
  what = c("all", "validation", "training"),
  digits = 4,
  print.gap = 2,
  ...
)

classify(object, ...)

\method{classify}{sopls}(object, classes, newdata, ncomp, LQ = "LDA", ...)

\method{R2}{sopls}(object, estimate, newdata, ncomp = "all", individual = FALSE, ...)

\method{RMSEP}{sopls}(object, estimate, newdata, ncomp = "all", individual = FALSE, ...)
}
\arguments{
\item{object}{A \code{sopls} object.}

\item{newdata}{Optional new data with the same types of predictor blocks as the ones used for fitting the object.}

\item{ncomp}{An \code{integer} giving the number of components to apply.}

\item{comps}{An \code{integer} vector giving the exact components to apply.}

\item{type}{A \code{character} for \code{predict} indicating if responses or scores should be predicted (default = "response", or "scores"), for \code{summary} indicating which type of explained variance to compute (default = "train", alternative = "CV").}

\item{na.action}{Function determining what to do with missing values in \code{newdata}.}

\item{...}{Additional arguments. Currently not implemented.}

\item{intercept}{A \code{logical} indicating if coefficients for the intercept should be included (default = FALSE).}

\item{x}{A \code{sopls} object.}

\item{what}{A \code{character} indicating if summary should include all, validation or training.}

\item{digits}{The number of digits used for printing.}

\item{print.gap}{Gap between columns when printing.}

\item{classes}{A \code{character} vector of class labels.}

\item{LQ}{A \code{character} indicating if 'max' (maximum score value), 'lda' or 'qda' should be used when classifying.}

\item{estimate}{A \code{character} indicating if 'train', 'CV' or 'test' results should be displayed.}

\item{individual}{A \code{logical} indicating if results for individual responses should be displayed.}

\item{block}{An \code{integer} indicating which block to use.}
}
\value{
Returns depend on method used, e.g. \code{predict.sopls} returns predicted responses
or scores depending on inputs, \code{coef.sopls} return regression coefficients.
}
\description{
Standard result functions for SO-PLS (\code{\link{sopls}}).
}
\examples{
data(potato)
mod <- sopls(Sensory[,1] ~ ., data = potato[c(1:3,9)], ncomp = 5, subset = 1:20)
testset <- potato[-(1:20),]; testset$Sensory <- testset$Sensory[,1,drop=FALSE]
predict(mod, testset, comps=c(2,1,2))
#dim(coef(mod, ncomp=5)) # <variables x responses x components>
print(mod)
summary(mod)

}
\references{
Jørgensen K, Mevik BH, Næs T. Combining designed experiments with several blocks of spectroscopic data. Chemometr Intell Lab Syst. 2007;88(2): 154–166.
}
\seealso{
Overviews of available methods, \code{\link{multiblock}}, and methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
