% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sopls_plot.R
\name{sopls_plot}
\alias{sopls_plot}
\alias{loadings.sopls}
\alias{scores.sopls}
\alias{loadingplot.sopls}
\alias{scoreplot.sopls}
\title{Scores, loadings and plots for sopls objects}
\usage{
\method{loadings}{sopls}(object, ncomp = "all", block = 1, ...)

\method{scores}{sopls}(object, ncomp = "all", block = 1, ...)

\method{scoreplot}{sopls}(
  object,
  comps = 1:2,
  ncomp = "all",
  block = 1,
  labels,
  identify = FALSE,
  type = "p",
  xlab,
  ylab,
  ...
)

\method{loadingplot}{sopls}(
  object,
  comps = 1:2,
  ncomp = "all",
  block = 1,
  scatter = TRUE,
  labels,
  identify = FALSE,
  type,
  lty,
  lwd = NULL,
  pch,
  cex = NULL,
  col,
  legendpos,
  xlab,
  ylab,
  pretty.xlabels = TRUE,
  xlim,
  ...
)
}
\arguments{
\item{object}{\code{sopls} object}

\item{ncomp}{\code{integer} vector giving components from all blocks before \code{block} (see next argument).}

\item{block}{\code{integer} indicating which block to extract components from.}

\item{...}{further arguments sent to the underlying plot function(s)}

\item{comps}{\code{integer} vector giving components, within block, to plot.}

\item{labels}{\code{character} indicating if "names" or "numbers" should be plot symbols (optional).}

\item{identify}{\code{logical} for activating \code{identify} to interactively identify points.}

\item{type}{\code{character} for selecting type of plot to make. Defaults to "p" (points) for scatter plots and "l" (lines) for line plots.}

\item{xlab}{\code{character} text for x labels.}

\item{ylab}{\code{character} text for y labels.}

\item{scatter}{\code{logical} indicating if a scatterplot of loadings should be made (default = TRUE).}

\item{lty}{Vector of line type specifications (see \code{\link{par}} for details).}

\item{lwd}{\code{numeric} vector of line width specifications.}

\item{pch}{Vector of point specifications (see \code{\link{points}} for details).}

\item{cex}{\code{numeric} vector of plot size expansions (see \code{\link{par}} for details).}

\item{col}{\code{integer} vector of symbol/line colours (see \code{\link{par}} for details).}

\item{legendpos}{\code{character} indicating legend position (if \code{scatter} is FALSE), e.g. \code{legendpos = "topright"}.}

\item{pretty.xlabels}{\code{logical} indicating if xlabels should be more nicely plotted (default = TRUE).}

\item{xlim}{\code{numeric} vector of length two, with the x limits of the plot (optional).}
}
\value{
The functions return whatever the underlying plot function (or identify) returns.
}
\description{
Adaptation of \code{scoreplot} from package \code{pls} for \code{sopls} objects.
}
\examples{
data(potato)
so <- sopls(Sensory ~ Chemical + Compression + NIRraw, data=potato, ncomp=c(5,5,5))

# Loadings
loadings(so, ncomp=c(3), block=2)[, 1:3]

# Scores
scores(so, block=1)[, 1:4]

# Default plot from first block
scoreplot(so)

# Second block with names
scoreplot(so, ncomp=c(3), block=2, labels="names")

# Scatterplot matrix
scoreplot(so, ncomp=c(3,2), block=3, comps=1:3)

# Default plot from first block
loadingplot(so, scatter=TRUE)

# Second block with names
loadingplot(so, ncomp=c(3), block=2, labels="names", scatter=TRUE)

# Scatterplot matrix
loadingplot(so, ncomp=c(3,2), block=3, comps=1:3, scatter=TRUE)
}
