% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised.R
\name{supervised}
\alias{supervised}
\alias{mbpls}
\alias{mbra}
\title{Supervised Multiblock Methods}
\usage{
mbpls(X, Y, ncomp = 1, scale = FALSE, ...)

mbra(X, Y, ncomp = 1, ...)
}
\description{
Collection of supervised multiblock methods:
\itemize{
\item MB-PLS - Multiblock Partial Least Squares (\code{mbpls})
\item SO-PLS - Sequential and Orthgonalized PLS (\code{\link{sopls}})
\item ROSA - Response Oriented Sequential Alternation (\code{\link{rosa}})
\item Multiblock Redundancy Analysis (\code{mbra})
}
}
