% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_results.R
\name{supervised_results}
\alias{supervised_results}
\alias{print.mbpls}
\alias{summary.mbpls}
\alias{blockLoadings.mbpls}
\alias{blockScores.mbpls}
\alias{blockLoadings}
\alias{blockScores}
\alias{print.mbrda}
\title{Result functions for Supervised Multiblock Methods}
\usage{
\method{print}{mbpls}(x, ...)

\method{summary}{mbpls}(object, ...)

blockLoadings(object, ...)

\method{blockLoadings}{mbpls}(object, block = 1, ...)

blockScores(object, ...)

\method{blockScores}{mbpls}(object, block = 1, ...)

\method{print}{mbrda}(x, ...)
}
\arguments{
\item{x}{\code{object} returned from \code{mbpls} or \code{mbrda}.}

\item{...}{additional arguments (not used).}

\item{object}{\code{object} returned from \code{mbpls} or \code{mbrda}.}

\item{block}{\code{integer} or \code{character} identifying a block when used block-wise.}
}
\description{
Result functions for Supervised Multiblock Methods
}
\examples{
data(potato)
mb <- mbpls(potato[c('Chemical','Compression')], potato[['Sensory']], ncomp = 5)
print(mb)

Tb1 <- blockScores(mb)
plot(Tb1)

}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
