% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{unsupervised}
\alias{unsupervised}
\alias{sca}
\alias{gca}
\alias{mfa}
\alias{pcagca}
\alias{disco}
\alias{jive}
\alias{statis}
\alias{hogsvd}
\alias{hpca}
\alias{mcoa}
\alias{gpa}
\title{Unsupervised Multiblock Methods}
\usage{
sca(X, ncomp = 2, scale = FALSE, samplelinked = "auto", ...)

gca(X, ncomp = 2, svd = TRUE, tol = 10^-12, corrs = TRUE, ...)

gpa(X, graph = FALSE, ...)

mfa(X, type = rep("c", length(X)), graph = FALSE, ...)

pcagca(
  X,
  commons = 2,
  auto = TRUE,
  auto.par = list(explVarLim = 40, rLim = 0.8),
  manual.par = list(ncomp = 0, ncommon = 0),
  tol = 10^-12
)

disco(X, ncomp = 2, ...)

hpca(X, ncomp = 2, scale = FALSE, verbose = FALSE, ...)

mcoa(X, ncomp = 2, scale = FALSE, verbose = FALSE, ...)

jive(X, ...)

statis(X, ncomp = 3, scannf = FALSE, tol = 1e-07, ...)

hogsvd(X)
}
\description{
Collection of unsupervised multiblock methods:
\itemize{
\item SCA - Simultaneous Component Analysis (\code{sca})
\item GCA - Generalized Canonical Analysis (\code{gca})
\item GPA - Generalized Procrustes Analysis (\code{gpa})
\item MFA - Multiple Factor Analysis (\code{mfa})
\item PCA-GCA (\code{pcagca})
\item DISCO - Distinctive and Common Components with SCA (\code{disco})
\item HPCA - Hierarchical Principal component analysis (\code{hpca})
\item MCOA - Multiple Co-Inertia Analysis (\code{mcoa})
\item JIVE - Joint and Individual Variation Explained (\code{jive})
\item STATIS - Structuration des Tableaux à Trois Indices de la Statistique (\code{statis})
\item HOGSVD - Higher Order Generalized SVD (\code{hogsvd})
}
}
\details{
Original documentation of STATIS: \link[ade4]{statis}.
JIVE, STATIS and HOGSVD assume variable linked matrices/data.frames, while SCA handles both links.
}
\examples{
# Object linked data
potList <- as.list(potato[c(1,2,9)])
pot.sca    <- sca(potList)
pot.gca    <- gca(potList)
pot.gpa    <- gpa(potList)
pot.mfa    <- mfa(potList)
pot.pcagca <- pcagca(potList)
pot.disco  <- disco(potList)
pot.hpca   <- hpca(potList)
pot.mcoa   <- mcoa(potList)

# Variable linked data
candyList <- lapply(1:nlevels(candies$candy),function(x)candies$assessment[candies$candy==x,])
can.jive   <- jive(candyList)
can.statis <- statis(candyList)
can.hogsvd <- hogsvd(candyList)
plot(can.statis$statis)
}
