% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_results.R
\name{unsupervised_results}
\alias{unsupervised_results}
\alias{scores.multiblock}
\alias{print.sca}
\alias{loadings.multiblock}
\alias{print.multiblock}
\alias{summary.multiblock}
\title{Result functions for Unsupervised Multiblock Methods}
\usage{
\method{scores}{multiblock}(object, block = 0, ...)

\method{loadings}{multiblock}(object, block = 0, ...)

\method{print}{multiblock}(x, ...)

\method{summary}{multiblock}(object, ...)
}
\description{
Result functions for Unsupervised Multiblock Methods
}
\examples{
data(wine)
sc <- sca(wine[c('Smell at rest', 'View', 'Smell after shaking')], ncomp = 4)
print(sc)

}
