% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic}
\alias{basic}
\alias{pca}
\title{Single- and Two-Block Methods}
\usage{
pca(X, scale = FALSE, ncomp = 1, ...)
}
\arguments{
\item{X}{\code{matrix} of input data.}

\item{scale}{\code{logical} indicating if variables should be standardised (default=FALSE).}

\item{ncomp}{\code{integer} number of principal components to return.}

\item{...}{additional arguments to \code{pls:pcr}.}
}
\value{
\code{multiblock} object with scores, loadings, mean X values and explained variances.
}
\description{
This documentation covers a range of single- and two-block methods. In particular:
\itemize{
\item PCA - Principal Component Analysis (\code{pca})
\item PCR - Principal Component Regression (\code{pcr})
\item PLSR - Partial Least Squares Regression (\code{plsr})
\item CCA - Canonical Correlation Analysis (\code{cca})
\item IFA - Interbattery Factor Analysis (\code{ifa})
\item GSVD - Generalized SVD (\code{gsvd})
}
}
\examples{
data(potato)
X <- potato$Chemical
y <- potato$Sensory[,1,drop=FALSE]

pca.pot  <- pca(X, ncomp = 2)
pcr.pot  <- pcr(y ~ X, ncomp = 2)
pls.pot  <- plsr(y ~ X, ncomp = 2)
cca.pot  <- cca(potato[1:2])
ifa.pot  <- ifa(potato[1:2])
gsvd.pot <- gsvd(lapply(potato[3:4], t))

}
\references{
\itemize{
\item PCA: Pearson, K. (1901) On lines and planes of closest fit to points in space. Philosophical Magazine, 2, 559–572.
\item PLS: Wold, S., Ruhe, A., Wold, H., and Dunn, W.J. (1984) The collinearity problem in linear regression. the partial least squares (PLS) approach to generalized inverses. SIAM Journal of Scientific and Statistical Computing, 5, 735–743.
\item CCA: Hotelling, H. (1936) Relations between two sets of variates. Biometrika, 28, 321–377.
\item IFA: Tucker, L. R. (1958). An inter-battery method of factor analysis. Psychometrika, 23(2), 111-136.
\item GSVD: Van Loan, C. (1976) Generalizing the singular value decomposition. SIAM Journal on Numerical Analysis, 13, 76–83.
}

Pearson, K. (1901) On lines and planes of closest fit to points in space. Philosophical Magazine, 2, 559–572.
}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
NULL
Principal Component Analysis - PCA

Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
