% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{import}
\name{basic}
\alias{basic}
\alias{pca}
\alias{pcr}
\alias{plsr}
\alias{cca}
\alias{gsvd}
\alias{ifa}
\alias{reexports}
\title{Single- and Two-Block Methods}
\usage{
pca(X, scale = FALSE, ncomp = 1, ...)

cca(X)

ifa(X, ncomp = 1, scale = FALSE, verbose = FALSE, ...)

gsvd(X)
}
\description{

}
\examples{
data(potato)
X <- potato$Chemical
y <- potato$Sensory[,1,drop=FALSE]

pca.pot  <- pca(X, ncomp = 2)
pcr.pot  <- pcr(y ~ X, ncomp = 2)
pls.pot  <- plsr(y ~ X, ncomp = 2)
cca.pot  <- cca(potato[1:2])
ifa.pot  <- ifa(potato[1:2])
gsvd.pot <- gsvd(lapply(potato[3:4], t))

}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pls}{\code{\link[pls:mvr]{pcr}}, \code{\link[pls:mvr]{plsr}}}
}}

