% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{import}
\name{basic}
\alias{basic}
\alias{pca}
\alias{pcr}
\alias{plsr}
\alias{ifa}
\alias{reexports}
\alias{cca}
\title{Single- and Two-Block Methods}
\usage{
pca(X, scale = FALSE, ncomp = 1, ...)

cca(X)

ifa(X, ncomp = 1, scale = FALSE, verbose = FALSE, ...)
}
\description{
This documentation covers a range of single- and two-block methods. In particular:
\itemize{
\item Principal Component Analysis (\code{pca})
\item Principal Component Regression (\code{pcr})
\item Partial Least Squares Regression (\code{plsr})
\item Canonical Correlation Analysis (\code{cca})
\item Interbattery Factor Analysis (\code{ifa})
}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pls}{\code{\link[pls:mvr]{pcr}}, \code{\link[pls:mvr]{plsr}}}
}}

