% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compnames}
\alias{compnames}
\title{Vector of component names}
\usage{
compnames(object, comps, explvar = FALSE, ...)
}
\arguments{
\item{object}{An object fitted using the multiblock package.}

\item{comps}{\code{integer} vector of components.}

\item{explvar}{\code{logical} indicating if explained variances should be included.}

\item{...}{Unused}
}
\value{
A \code{character} vector of component names.
}
\description{
Convenience function for creating a vector
of component names based on the dimensions the input object
(\code{matrix} or object having a \code{score} function).
}
\details{
This is a copy of \code{compnames} from the \code{pls} package to work with
\code{multiblock} objects.
}
