% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{gca}
\alias{gca}
\title{Generalized Canonical Analysis - GCA}
\usage{
gca(X, ncomp = 2, svd = TRUE, tol = 10^-12, corrs = TRUE, ...)
}
\arguments{
\item{X}{\code{list} of input blocks.}

\item{ncomp}{\code{integer} number of components to extract.}

\item{svd}{\code{logical} indicating if Singular Value Decomposition approach should be used (default=TRUE).}

\item{tol}{\code{numeric} tolerance for component inclusion (singular values).}

\item{corrs}{\code{logical} indicating if correlations should be calculated for RGCCA based approach.}

\item{...}{additional arguments for RGCCA approach.}
}
\description{
Generalized Canonical Analysis - GCA
}
\examples{
data(potato)
potList <- as.list(potato[c(1,2,9)])
pot.gca <- gca(potList)
plot(scores(pot.gca), labels="names")

}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
