% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{gsvd}
\alias{gsvd}
\title{Generalised Singular Value Decomposition}
\usage{
gsvd(X)
}
\arguments{
\item{X}{\code{matrix} of input data.}
}
\value{
\code{multiblock} object with associated with printing, scores, loadings, ...
}
\description{
Generalised Singular Value Decomposition
}
\references{
Van Loan, C. (1976) Generalizing the singular value decomposition. SIAM Journal on Numerical Analysis, 13, 76–83.
}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
