% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{mfa}
\alias{mfa}
\title{Multiple Factor Analysis - MFA}
\usage{
mfa(X, type = rep("c", length(X)), graph = FALSE, ...)
}
\arguments{
\item{X}{\code{list} of input blocks.}

\item{type}{\code{character} vector indicating block types, defaults to \code{rep("c", length(X))} for continuous values.}

\item{graph}{\code{logical} indicating if decomposition should be plotted.}

\item{...}{additional arguments for RGCCA approach.}
}
\description{
Multiple Factor Analysis - MFA
}
\examples{
data(potato)
potList <- as.list(potato[c(1,2,9)])
pot.mfa    <- mfa(potList)
if(interactive()){
  plot(pot.mfa$MFA)
}

}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
