% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_object.R, R/sopls_plot.R
\name{multiblock_object}
\alias{multiblock_object}
\alias{scores.multiblock}
\alias{loadings.multiblock}
\alias{print.multiblock}
\alias{summary.multiblock}
\alias{scoreplot.multiblock}
\alias{loadingplot.multiblock}
\alias{loadingweightplot}
\alias{biplot.multiblock}
\alias{biplot.sopls}
\title{Result Functions for Supervised and Multiblock Objects}
\usage{
\method{scores}{multiblock}(object, block = 0, ...)

\method{loadings}{multiblock}(object, block = 0, ...)

\method{print}{multiblock}(x, ...)

\method{summary}{multiblock}(object, ...)

\method{scoreplot}{multiblock}(
  object,
  comps = 1:2,
  block = 0,
  labels,
  identify = FALSE,
  type = "p",
  xlab,
  ylab,
  main,
  ...
)

\method{loadingplot}{multiblock}(
  object,
  comps = 1:2,
  block = 0,
  scatter = TRUE,
  labels,
  identify = FALSE,
  type,
  lty,
  lwd = NULL,
  pch,
  cex = NULL,
  col,
  legendpos,
  xlab,
  ylab,
  main,
  pretty.xlabels = TRUE,
  xlim,
  ...
)

loadingweightplot(object, main = "Loading weights", ...)

\method{biplot}{multiblock}(
  x,
  block = 0,
  comps = 1:2,
  which = c("x", "y", "scores", "loadings"),
  var.axes = FALSE,
  xlabs,
  ylabs,
  main,
  ...
)

\method{biplot}{sopls}(
  x,
  comps = 1:2,
  ncomp = "all",
  block = 1,
  which = c("x", "y", "scores", "loadings"),
  var.axes = FALSE,
  xlabs,
  ylabs,
  main,
  ...
)
}
\arguments{
\item{object}{\code{multiblock} object.}

\item{block}{\code{integer/character} for block selection.}

\item{...}{Not implemented.}

\item{x}{\code{multiblock} object.}

\item{comps}{\code{integer} vector giving components, within block, to plot.}

\item{labels}{\code{character} indicating if "names" or "numbers" should be plot symbols (optional).}

\item{identify}{\code{logical} for activating \code{identify} to interactively identify points.}

\item{type}{\code{character} for selecting type of plot to make. Defaults to "p" (points) for scatter plots and "l" (lines) for line plots.}

\item{xlab}{\code{character} text for x labels.}

\item{ylab}{\code{character} text for y labels.}

\item{main}{\code{character} text for main header.}

\item{scatter}{\code{logical} indicating if a scatterplot of loadings should be made (default = TRUE).}

\item{lty}{Vector of line type specifications (see \code{\link{par}} for details).}

\item{lwd}{\code{numeric} vector of line width specifications.}

\item{pch}{Vector of point specifications (see \code{\link{points}} for details).}

\item{cex}{\code{numeric} vector of plot size expansions (see \code{\link{par}} for details).}

\item{col}{\code{integer} vector of symbol/line colours (see \code{\link{par}} for details).}

\item{legendpos}{\code{character} indicating legend position (if \code{scatter} is FALSE), e.g. \code{legendpos = "topright"}.}

\item{pretty.xlabels}{\code{logical} indicating if xlabels should be more nicely plotted (default = TRUE).}

\item{xlim}{\code{numeric} vector of length two, with the x limits of the plot (optional).}
}
\description{
Result Functions for Supervised and Multiblock Objects
}
\examples{
data(wine)
sc <- sca(wine[c('Smell at rest', 'View', 'Smell after shaking')], ncomp = 4)
print(sc)
plot(loadings(sc, block = 1), labels = "names", scatter = TRUE)

}
\seealso{
Overviews of available methods, \code{\link{multiblock}}, and methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
