% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_indirect.R
\name{SO_TDI}
\alias{SO_TDI}
\alias{sopls_dirIndTotAddOver}
\alias{print.SO_TDI}
\alias{sopls_dirIndTotAddOver_multiple}
\alias{print.SO_TDI_multiple}
\title{Total, direct, indirect and additional effects in SO-PLS-PM.}
\usage{
sopls_dirIndTotAddOver(
  X,
  Y,
  comps,
  max_comps = min(sum(comps), 20),
  sel.comp = "opt",
  computeAdditional = FALSE,
  sequential = FALSE,
  B = NULL,
  k = 10,
  type = "consecutive"
)

\method{print}{SO_TDI}(x, showComp = TRUE, heading = "SO-PLS path effects", digits = 2, ...)

sopls_dirIndTotAddOver_multiple(
  X,
  X_names,
  comps,
  max_comps = min(sum(comps), 20),
  sel.comp = "opt",
  computeAdditional = FALSE,
  sequential = FALSE,
  B = NULL,
  k = 10,
  type = "consecutive"
)

\method{print}{SO_TDI_multiple}(x, heading = "SO-PLS path effects", digits = 2, ...)
}
\arguments{
\item{X}{A \code{list} of input blocks (of type \code{matrix}).}

\item{Y}{A \code{matrix} of response(s).}

\item{comps}{An \code{integer} vector giving the number of components per block.}

\item{max_comps}{Maximum total number of components.}

\item{sel.comp}{A \code{character} or \code{integer} vector indicating the type ("opt" - minimum error / "chi" - chi-squared reduced) or number of components in selections.}

\item{computeAdditional}{A \code{logical} indicating if additional components should be computed.}

\item{sequential}{A \code{logical} indicating if sequential component optimization should be applied.}

\item{B}{An \code{integer} giving the number of bootstrap replicates for variation estimation.}

\item{k}{An \code{integer} indicating number of cross-validation segments (default = 10).}

\item{type}{A \code{character} for selecting type of cross-validation segments (default = "consecutive").}

\item{x}{An object of type \code{SO_TDI}.}

\item{showComp}{A \code{logical} indicating if components should be shown in print (default = TRUE).}

\item{heading}{A \code{character} giving the heading of the print.}

\item{digits}{An \code{integer} for selecting number of digits in print.}

\item{...}{Not implemented}

\item{X_names}{A \code{list} of names for the X blocks.}
}
\value{
An object of type \code{SO_TDI} containing total, direct and indirect effects, plus
possibly additional effects and standard deviations.
}
\description{
Total, direct, indirect and additional effects in SO-PLS-PM.
}
