% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised.R
\name{supervised}
\alias{supervised}
\alias{mbpls}
\alias{mbrda}
\title{Supervised Multiblock Methods}
\usage{
mbpls(X, Y, ncomp = 1, scale = FALSE, ...)

mbrda(X, Y, ncomp = 1, ...)
}
\description{
Collection of supervised multiblock methods:
\itemize{
\item MB-PLS - Multiblock Partial Least Squares (\code{mbpls})
\item SO-PLS - Sequential and Orthogonalized PLS (\code{\link{sopls}})
\item ROSA - Response Oriented Sequential Alternation (\code{\link{rosa}})
\item mbRDA - Multiblock Redundancy Analysis (\code{mbrda})
}
}
\examples{
data(potato)
mb <- mbpls(potato[c('Chemical','Compression')], potato[['Sensory']], ncomp = 5)
print(mb)

# Convert data.frame with AsIs objects to list of matrices
potatoList <- lapply(potato, unclass)
mbr <- mbrda(potatoList[c('Chemical','Compression')], potatoList[['Sensory']], ncomp = 10)
print(mbr)
scoreplot(mbr)

TODO: 
* mbRDA to formula interface
* MBPLS to formula interface
* sparse MBPLS (ST-type, Lenth type)
* PO-PLS

}
