% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_results.R
\name{supervised_results}
\alias{supervised_results}
\alias{print.mbpls}
\alias{summary.mbpls}
\alias{blockLoadings.mbpls}
\alias{blockScores.mbpls}
\alias{blockLoadings}
\alias{blockScores}
\title{Result functions for Supervised Multiblock Methods}
\usage{
\method{print}{mbpls}(x, ...)

\method{summary}{mbpls}(object, ...)

blockLoadings(object, ...)

\method{blockLoadings}{mbpls}(object, block = 1, ...)

blockScores(object, ...)

\method{blockScores}{mbpls}(object, block = 1, ...)
}
\description{
Result functions for Supervised Multiblock Methods
}
\examples{
data(potato)
mb <- mbpls(potato[c('Chemical','Compression')], potato[['Sensory']], ncomp = 5)
print(mb)

Tb1 <- blockScores(mb)
plot(Tb1)

}
