% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{unsupervised}
\alias{unsupervised}
\alias{hpca}
\alias{gca}
\alias{sca}
\alias{mfa}
\alias{mcoa}
\alias{pcagca}
\alias{disco}
\alias{statis}
\alias{hogsvd}
\title{Unsupervised Multiblock Methods}
\usage{
hpca(X, ncomp = 1, scale = FALSE, verbose = FALSE, ...)

gca(X, ncomp = 1, svd = TRUE, tol = 10^-12, corrs = TRUE, ...)

sca(X, ncomp = 1, scale = FALSE, samplelinked = "auto", ...)

mfa(X, groupType = rep("n", length(X)), groupName = NULL, ...)

mcoa(X, ncomp = 1, scale = FALSE, verbose = FALSE, ...)

pcagca(
  X,
  commons = 2,
  auto = TRUE,
  auto.par = list(explVarLim = 40, rLim = 0.8),
  manual.par = list(ncomp = 0, ncommon = 0),
  tol = 10^-12
)

disco(X, ncomp, ...)

statis(X, ncomp = 3, scannf = FALSE, tol = 1e-07, ...)

hogsvd(X, ncomp = 3)
}
\description{
Collection of unsupervised multiblock methods:
\itemize{
\item Hierarchical Principal component analysis (\code{hpca})
\item Generalized Canonical Analysis (\code{gca})
\item Simultaneous Component Analysis (\code{sca})
\item Higher Order Generalized SVD (\code{hogsvd})
\item Multiple Factor Analysis (\code{mfa})
\item Multiple Co-Inertia Analysis (\code{mcoa})
\item PCA-GCA (\code{pcagca})
\item Distinctive and Common Components with SCA (\code{disco})
\item STATIS - Structuration des Tableaux A Trois Indices de la Statistique (\code{statis})
}

Original documentation of STATIS: \link[ade4]{statis}.
}
