% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{unsupervised}
\alias{unsupervised}
\title{Unsupervised Multiblock Methods}
\description{

}
\details{
Original documentation of STATIS: \link[ade4]{statis}.
JIVE, STATIS and HOGSVD assume variable linked matrices/data.frames, while SCA handles both links.
}
\examples{
# Object linked data
data(potato)
potList <- as.list(potato[c(1,2,9)])
pot.sca    <- sca(potList)

# Variable linked data
data(candies)
candyList <- lapply(1:nlevels(candies$candy),function(x)candies$assessment[candies$candy==x,])
can.statis <- statis(candyList)
plot(can.statis$statis)
}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
