% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{unsupervised}
\alias{unsupervised}
\title{Unsupervised Multiblock Methods}
\description{
Collection of unsupervised multiblock methods:
\itemize{
\item SCA - Simultaneous Component Analysis (\code{sca})
\item GCA - Generalized Canonical Analysis (\code{gca})
\item GPA - Generalized Procrustes Analysis (\code{gpa})
\item MFA - Multiple Factor Analysis (\code{mfa})
\item PCA-GCA (\code{pcagca})
\item DISCO - Distinctive and Common Components with SCA (\code{disco})
\item HPCA - Hierarchical Principal component analysis (\code{hpca})
\item MCOA - Multiple Co-Inertia Analysis (\code{mcoa})
\item JIVE - Joint and Individual Variation Explained (\code{jive})
\item STATIS - Structuration des Tableaux à Trois Indices de la Statistique (\code{statis})
\item HOGSVD - Higher Order Generalized SVD (\code{hogsvd})
}
}
\details{
Original documentation of STATIS: \link[ade4]{statis}.
JIVE, STATIS and HOGSVD assume variable linked matrices/data.frames, while SCA handles both links.
}
\examples{
# Object linked data
data(potato)
potList <- as.list(potato[c(1,2,9)])
pot.sca    <- sca(potList)

# Variable linked data
data(candies)
candyList <- lapply(1:nlevels(candies$candy),function(x)candies$assessment[candies$candy==x,])
can.statis <- statis(candyList)
plot(can.statis$statis)
}
\seealso{
Overviews of available methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{asca}}, \code{\link{supervised}} and \code{\link{complex}}.
}
