% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{unsupervised}
\alias{unsupervised}
\alias{sca}
\alias{gca}
\alias{mfa}
\alias{pcagca}
\alias{disco}
\alias{jive}
\alias{statis}
\alias{hogsvd}
\alias{hpca}
\alias{mcoa}
\title{Unsupervised Multiblock Methods}
\usage{
sca(X, ncomp = 1, scale = FALSE, samplelinked = "auto", ...)

gca(X, ncomp = 1, svd = TRUE, tol = 10^-12, corrs = TRUE, ...)

mfa(X, groupType = rep("c", length(X)), groupName = NULL, ...)

pcagca(
  X,
  commons = 2,
  auto = TRUE,
  auto.par = list(explVarLim = 40, rLim = 0.8),
  manual.par = list(ncomp = 0, ncommon = 0),
  tol = 10^-12
)

disco(X, ncomp = 2, ...)

jive(X, ...)

statis(X, ncomp = 3, scannf = FALSE, tol = 1e-07, ...)

hogsvd(X, ncomp = unlist(lapply(X, function(x) min(dim(x)))) - 1)

hpca(X, ncomp = 1, scale = FALSE, verbose = FALSE, ...)

mcoa(X, ncomp = 1, scale = FALSE, verbose = FALSE, ...)
}
\description{
Collection of unsupervised multiblock methods:
\itemize{
\item SCA - Simultaneous Component Analysis (\code{sca})
\item GCA - Generalized Canonical Analysis (\code{gca})
\item MFA - Multiple Factor Analysis (\code{mfa})
\item PCA-GCA (\code{pcagca})
\item DISCO - Distinctive and Common Components with SCA (\code{disco})
\item JIVE - Joint and Invidual Variation Explained (\code{jive})
\item STATIS - Structuration des Tableaux A Trois Indices de la Statistique (\code{statis})
\item HOGSVD - Higher Order Generalized SVD (\code{hogsvd})
\item HPCA - Hierarchical Principal component analysis (\code{hpca})
\item MCOA - Multiple Co-Inertia Analysis (\code{mcoa})
}
}
\details{
Original documentation of STATIS: \link[ade4]{statis}.
}
\examples{
potList <- as.list(potato[1:3])
pot.sca    <- sca(potList)
pot.gca    <- gca(potList)
pot.mfa    <- mfa(potList)
pot.pcagca <- pcagca(potList)
pot.disco  <- disco(potList)
pot.jive   <- jive(potList)
pot.hogsvd <- hogsvd(potList)
summary(pot.jive)
}
