% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_results.R
\name{unsupervised_results}
\alias{unsupervised_results}
\alias{print.sca}
\alias{blockScores.sca}
\alias{blockLoadings.sca}
\title{Result functions for Unsupervised Multiblock Methods}
\usage{
\method{print}{sca}(x, ...)

\method{blockScores}{sca}(object, block = 1, ...)

\method{blockLoadings}{sca}(object, block = 1, ...)
}
\description{
Result functions for Unsupervised Multiblock Methods
}
\examples{
data(wine)
sc <- sca(wine[c('Smell at rest', 'View', 'Smell after shaking')], ncomp = 4)
print(sc)

}
